(Simplified File Structure)
My IDE/OS doesn't recognize a lot of the relative imports I have in the project -- namely, because it doesn't register `__ini__.py` files as anything besides plain text.

To avoid this problem in the future, I'm going to simplify the file structure and import system.

Changes Summary:
- changed imports on rx.py cookbook
- moved simple modules to files
