--- hcm/dynamical/__init__.py
@@ -1,1 +1,0 @@
-from hcm.dynamical.lorenz import lorenz
