--- hcm/io/__init__.py
@@ -1,2 +1,0 @@
-from hcm.io.wav import wav_read, wav_write
-from hcm.io.channel import add_channels, append
