--- hcm/io/channel.py
@@ -1,9 +1,0 @@
-import numpy as np
-
-
-def add_channels(signals):
-    return np.array(np.vstack(signals)).T
-
-
-def append(*signals):
-    return np.array(np.hstack(signals))
