+++ hcm/signal/__init__.py
@@ -1,2 +1,2 @@
-from hcm.signal.signal import time
+from hcm.signal.time import time
 from hcm.signal.vco import VCO
