+++ hcm/signal/noise.py
