--- hcm/signal/noise/__init__.py
@@ -1,1 +1,0 @@
-from hcm.signal.noise.noise import brownian_noise, white_noise
