+++ hcm/signal/osc.py
