--- hcm/signal/osc/__init__.py
@@ -1,1 +1,0 @@
-from hcm.signal.osc.oscillator import sine, triangle, square
