+++ hcm/signal/time.py
