--- hcm/ts/__init__.py
@@ -1,1 +1,0 @@
-from hcm.ts.timeseries import normalize, AM, sample_and_hold
